Added adapter reload button
Tried and failed to automatically refresh the adapters.
Running a while loop inside the existing powershell thread locks up
the program. I am going to try to revist this at a later time.

Clicking the reload button will clear the list box then get the current
adapters that are up

Signed-off-by: Ryan Huntington <rhuntington719@gmail.com>