Add New IP to Adapter Enabled new IP button which allows user to add IPs to an adapter
It needs further refining to better check for an existing IP on an
adapter. I currently have a general garbage error handler that states
that the IP already exists or things went sideways.

In the next major update, I want to add hot buttons to quickly set IPs
for common scenarios such as Axis and Genetec.

Signed-off-by: Darth Killian <rhuntington719@gmail.com>